+++ backend/src/main/kotlin/com/vdsirotkin/telegram/mystickersbot/service/StickerPackManagementService.kt
@@ -20,7 +20,7 @@ class StickerPackManagementService(
                                          sticker: File,
                                          entity: UserEntity,
                                          emoji: String? = null) {
-        bot.executeAsync(AddStickerToSet(chatId.toInt(), entity.normalPackName, sticker, emoji ?: "🙂"))
+        bot.executeAsync(AddStickerToSet(chatId, entity.normalPackName, sticker, emoji ?: "🙂"))
     }
 
     suspend fun createNewPack(bot: TelegramBot,
@@ -29,7 +29,7 @@ class StickerPackManagementService(
                                       entity: UserEntity,
                                       emoji: String? = null,
                                       maskPosition: MaskPosition? = null) {
-        bot.executeAsync(CreateNewStickerSet(chatId.toInt(), entity.normalPackName, "Your stickers - @${props.username}", sticker, emoji
+        bot.executeAsync(CreateNewStickerSet(chatId, entity.normalPackName, "Your stickers - @${props.username}", sticker, emoji
                 ?: "🙂")
                 .apply {
                     if (maskPosition != null) {
